Fix instances of -Wsign-compare in mcman

+++ iop/memorycard/mcman/src/ps1mc_fio.c
@@ -152,7 +152,7 @@ int mcman_open1(int port, int slot, char *filename, int flags)
 	for (i = 0; i < MAX_FDHANDLES; i++) {
 		if ((mcman_fdhandles[i].status != 0) && \
 			(mcman_fdhandles[i].port == port) && (mcman_fdhandles[i].slot == slot) && \
-			 	(mcman_fdhandles[i].freeclink == r) && (fh->wrflag != 0))
+			 	(mcman_fdhandles[i].freeclink == (u32)r) && (fh->wrflag != 0))
 			return sceMcResDeniedPermit;
 	}
 
@@ -256,7 +256,7 @@ int mcman_read1(int fd, void *buffer, int nbyte)
 	if (fh->position >= fh->filesize)
 		return 0;
 
-	if (nbyte >= (fh->filesize - fh->position))
+	if (nbyte >= (int)(fh->filesize - fh->position))
 		nbyte = fh->filesize - fh->position;
 
 	rpos = 0;
